+++ src/ui/gameover/FinalScoreModal.tsx
@@ -1,12 +1,20 @@
 import React, { useState, useEffect } from "react";
-import { Dialog, DialogTitle, Button } from "@material-ui/core";
+import {
+  Dialog,
+  DialogTitle,
+  Button,
+  Card,
+  CardContent,
+} from "@material-ui/core";
 import styled from "styled-components";
 import { useHistory } from "react-router-dom";
-import { ScorePad } from "../score/ScorePad";
 import { useGameState } from "../GameContext";
 import { getLeaders } from "../../game/entities/score.utils";
 import { PlayerID } from "../../game/entities/players";
 import { getPlayerName } from "../../game/entities/players.utils";
+import { ScoreContainer } from "../score/ScoreContainer";
+import { ShareScore } from "./ShareScore";
+import { ExternalLink } from "../components/ExternalLink";
 
 export const FinalScoreModal: React.FC = () => {
   const [showModal, setShowModal] = useState(false);
@@ -34,7 +42,11 @@ const FinalScoreModalContent: React.FC<FinalScoreModalContentProps> = ({
   winnerIDs,
 }) => {
   const history = useHistory();
-  const { gameMetadata } = useGameState();
+  const {
+    gameMetadata,
+    wizardState: { scorePad },
+    ctx: { gameover },
+  } = useGameState();
 
   const winnerNames = winnerIDs.map((winnerID) =>
     getPlayerName(winnerID, gameMetadata ?? [])
@@ -45,8 +57,37 @@ const FinalScoreModalContent: React.FC<FinalScoreModalContentProps> = ({
         {winnerNames.join("&")}{" "}
         {winnerNames.length > 1 ? "gewinnen" : "gewinnt"}!
       </DialogTitle>
-      <ScorePad />
-      <ActionContainer>
+      <SectionContainer>
+        <ScoreContainer />
+      </SectionContainer>
+
+      <SectionContainer>
+        <Card>
+          <CardContent>
+            <h3>Hilf mit, Wizzzzard zu verbessern!</h3>
+            <p>
+              Mit deinem Feedback können wir Wizzzzard noch besser machen. Hast
+              du einen Fehler entdeckt, wünschst du dir neue Funktionen oder
+              möchtest uns etwas anderes mitteilen?{" "}
+              <ExternalLink href={process.env.FEEDBACK_FORM}>
+                Dann gib uns Feedback!
+              </ExternalLink>
+            </p>
+          </CardContent>
+        </Card>
+      </SectionContainer>
+
+      <SectionContainer>
+        <ShareScore
+          finalResult={{
+            date: new Date(gameover),
+            playerNames: (gameMetadata ?? []).map(({ name }) => name),
+            scorePad,
+          }}
+        />
+      </SectionContainer>
+
+      <SectionContainer>
         <Button
           color="primary"
           variant="contained"
@@ -54,7 +95,7 @@ const FinalScoreModalContent: React.FC<FinalScoreModalContentProps> = ({
         >
           Spiel schließen
         </Button>
-      </ActionContainer>
+      </SectionContainer>
     </Container>
   );
 };
@@ -63,6 +104,6 @@ const Container = styled.div`
   margin: 25px;
 `;
 
-const ActionContainer = styled.div`
+const SectionContainer = styled.div`
   margin: 15px 0;
 `;