+++ src/ui/lobby/CreateGame.tsx
@@ -1,10 +1,9 @@
 import React, { useState } from "react";
-import { useHistory, Link as RouterLink } from "react-router-dom";
+import { useHistory } from "react-router-dom";
 import {
   Select,
   MenuItem,
   Button,
-  Link,
   Checkbox,
   FormControlLabel,
 } from "@material-ui/core";
@@ -18,6 +17,8 @@ export const CreateGame: React.FC = () => {
   const history = useHistory();
   const [numPlayers, setNumPlayers] = useState<NumPlayers>(3);
   const [tournamentMode, setTournamentMode] = useState(false);
+  const [inspectPreviousTrick, setInspectPreviousTrick] = useState(false);
+
   return (
     <div>
       <h1>Starte ein Spiel</h1>
@@ -27,6 +28,7 @@ export const CreateGame: React.FC = () => {
             const gameID = await createGame(numPlayers, {
               config: {
                 tournamentMode,
+                inspectPreviousTrick,
               },
             });
             history.push(`/games/${gameID}`);
@@ -60,15 +62,25 @@ export const CreateGame: React.FC = () => {
             />
           </FieldContainer>
           <FieldContainer>
+            <FormControlLabel
+              label="Letzten Stich betrachten"
+              control={
+                <Checkbox
+                  checked={inspectPreviousTrick}
+                  onChange={(event) =>
+                    setInspectPreviousTrick(event.target.checked)
+                  }
+                />
+              }
+            />
+          </FieldContainer>
+          <FieldContainer>
             <Button type="submit" color="primary" variant="contained">
               Spiel starten
             </Button>
           </FieldContainer>
         </Form>
       </FormContainer>
-      <Link component={RouterLink} to="/games">
-        Alle Spiele anzeigen
-      </Link>
     </div>
   );
 };