+++ src/ui/lobby/LobbyRouter.tsx
@@ -2,15 +2,12 @@ import React, { useEffect } from "react";
 import { Switch, Route, Redirect, useHistory } from "react-router-dom";
 import { CreateGame } from "./CreateGame";
 import { GameContainer } from "./GameContainer";
-import { useProfileHeaderElement } from "../hooks/profile-header-element";
 import { ListGames } from "./ListGames";
-import { useFeedbackHeaderElement } from "../hooks/feedback-header-element";
 import { pageview } from "../../analytics";
+import { Profile } from "./Profile";
+import { FinalScore } from "./FinalScore";
 
 export const LobbyRouter: React.FC = () => {
-  useProfileHeaderElement();
-  useFeedbackHeaderElement();
-
   const history = useHistory();
   useEffect(() => {
     const unlisten = history.listen(pageview);
@@ -27,6 +24,12 @@ export const LobbyRouter: React.FC = () => {
       </Route>
       <Route path="/games/:gameID">
         <GameContainer />
+      </Route>
+      <Route path="/profile">
+        <Profile />
+      </Route>
+      <Route path="/score/:sharableFinalScore">
+        <FinalScore />
       </Route>
       <Route path="*">
         <Redirect to="/" />