+++ src/ui/score/ScorePad.tsx
@@ -8,44 +8,56 @@ import {
 } from "@material-ui/core";
 import range from "lodash/range";
 import styled from "styled-components";
-import { useGameState } from "../GameContext";
-import { getPlayerName, maxCards } from "../../game/entities/players.utils";
-import { PlayerID } from "../../game/entities/players";
 import { ScoreRow } from "./ScoreRow";
-import { RoundColCell, PlayerColCell } from "./Cells";
-import { ScoreCellProps } from "./ScoreCell";
-import { colors } from "../util/colors";
+import { RoundColCell, PlayerNameCell, ScoreCellProps } from "./Cells";
+import { ScorePad as ScorePadModel } from "../../game/entities/score";
+import { NumPlayers } from "../../game/entities/players";
+import { maxCards } from "../../game/entities/players.utils";
 
-export const ScorePad: React.FC = () => {
-  const {
-    wizardState: { scorePad, roundIndex, rounds, round, numPlayers },
-    gameMetadata,
-  } = useGameState();
-  const playerIDs = range(0, numPlayers) as PlayerID[];
-  const currentRound = rounds[roundIndex];
+export interface ScorePadProps {
+  playerNames: string[];
+  scorePad: ScorePadModel;
+  rounds: number[];
+  currentRound?: number;
+  currentRoundBids?: (number | undefined)[];
+}
+
+export const ScorePad: React.FC<ScorePadProps> = ({
+  playerNames,
+  scorePad,
+  rounds,
+  currentRound,
+  currentRoundBids,
+}) => {
+  const numPlayers = playerNames.length as NumPlayers;
   const allRounds = range(1, maxCards(numPlayers) + 1);
   const [colHighlighted, setColHighlighted] = useState(-1);
+  const [rowHighlighted, setRowHighlighted] = useState(-1);
+  const [highlightAll, setHighlightAll] = useState(false);
   return (
     <Container>
       <StyledTable>
         <Table size="small">
           <TableHead>
             <TableRow>
-              <RoundColCell />
-              {playerIDs.map((playerID) => (
+              <RoundColCell
+                onMouseEnter={() => setHighlightAll(true)}
+                onMouseLeave={() => setHighlightAll(false)}
+              />
+              {playerNames.map((playerName, i) => (
                 <PlayerNameCell
-                  key={playerID}
-                  $highlight={colHighlighted === playerID}
-                  onMouseEnter={() => setColHighlighted(playerID)}
+                  key={playerName}
+                  highlight={colHighlighted === i || highlightAll}
+                  onMouseEnter={() => setColHighlighted(i)}
                   onMouseLeave={() => setColHighlighted(-1)}
                 >
-                  {getPlayerName(playerID, gameMetadata)}
+                  {playerName}
                 </PlayerNameCell>
               ))}
             </TableRow>
           </TableHead>
           <TableBody>
-            {allRounds.map((numCards) => {
+            {allRounds.map((numCards, i) => {
               const roundScore = scorePad.find(
                 (scoreRow) => scoreRow.numCards === numCards
               );
@@ -55,19 +67,27 @@ export const ScorePad: React.FC = () => {
               );
               if (roundScore?.playerScores) {
                 playerScores = roundScore.playerScores;
-              } else if (round && isCurrentRound) {
-                playerScores = round.bids.map((bid) => ({
-                  bid: bid ?? undefined,
+              } else if (isCurrentRound && currentRoundBids) {
+                playerScores = currentRoundBids.map((bid) => ({
+                  bid,
                 }));
+              }
+              const skip = !rounds.includes(numCards);
+              let thisColHighlighted: number | "all" | undefined;
+              if (!skip) {
+                thisColHighlighted =
+                  highlightAll || rowHighlighted === i ? "all" : colHighlighted;
               }
               return (
                 <ScoreRow
                   numCards={numCards}
                   playerScores={playerScores}
-                  skip={!rounds.includes(numCards)}
+                  skip={skip}
                   current={isCurrentRound}
-                  colHighlighted={colHighlighted}
+                  colHighlighted={thisColHighlighted}
                   key={numCards}
+                  onMouseEnter={() => setRowHighlighted(i)}
+                  onMouseLeave={() => setRowHighlighted(-1)}
                 />
               );
             })}
@@ -94,12 +114,4 @@ const StyledTable = styled(TableContainer)`
     border: 1px solid black;
     border-width: 1px 2px;
   }
-`;
-
-const PlayerNameCell = styled(PlayerColCell)<{ $highlight: boolean }>`
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  background-color: ${({ $highlight }) =>
-    $highlight ? colors.yellow.light : "inherit"};
 `;