+++ src/ui/table/TableRound.tsx
@@ -1,14 +1,16 @@
 import React from "react";
 import styled from "styled-components";
 import { useGameState } from "../GameContext";
-import { Trick } from "./Trick";
+import { CurrentTrick } from "./CurrentTrick";
 import { BidRound } from "./BidRound";
 
 export const TableRound: React.FC = () => {
   const {
     wizardState: { trick, round },
   } = useGameState();
-  return <Container>{trick && round ? <Trick /> : <BidRound />}</Container>;
+  return (
+    <Container>{trick && round ? <CurrentTrick /> : <BidRound />}</Container>
+  );
 };
 
 const Container = styled.div`